+++ src/ems.h
@@ -34,19 +34,6 @@
 #define EMS_THERMOSTAT_WRITE_YES true
 #define EMS_THERMOSTAT_WRITE_NO false
 
-// ANSI Colors
-#define COLOR_RESET "\x1B[0m"
-#define COLOR_BLACK "\x1B[0;30m"
-#define COLOR_RED "\x1B[0;31m"
-#define COLOR_GREEN "\x1B[0;32m"
-#define COLOR_YELLOW "\x1B[0;33m"
-#define COLOR_BLUE "\x1B[0;34m"
-#define COLOR_MAGENTA "\x1B[0;35m"
-#define COLOR_CYAN "\x1B[0;36m"
-#define COLOR_WHITE "\x1B[0;37m"
-#define COLOR_BOLD_ON "\x1B[1m"
-#define COLOR_BOLD_OFF "\x1B[21m"
-
 // trigger settings to determine if hot tap water or the heating is active
 #define EMS_BOILER_BURNPOWER_TAPWATER 100
 #define EMS_BOILER_SELFLOWTEMP_HEATING 70