+++ src/my_config.h
@@ -16,8 +16,9 @@
 // MQTT general settings
 #define MQTT_TOPIC_START "start"
 #define MQTT_TOPIC_START_PAYLOAD "start"
-#define MQTT_WILL_TOPIC "status" // for last will & testament topic name
-#define MQTT_WILL_PAYLOAD "0"    // for last will & testament payload
+#define MQTT_WILL_TOPIC "status"            // for last will & testament topic name
+#define MQTT_WILL_ONLINE_PAYLOAD "online"   // for last will & testament payload
+#define MQTT_WILL_OFFLINE_PAYLOAD "offline" // for last will & testament payload
 #define MQTT_RETAIN false
 #define MQTT_KEEPALIVE 300
 #define MQTT_QOS 1
@@ -53,8 +54,9 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 // Set LED pin used for showing ems bus connection status. Solid is connected, Flashing is error
-// can be either the onboard LED on the ESP8266 (LED_BULLETIN) or external via an external pull-up LED (e.g. D1 on bbqkees' board)
-// can be enabled and disabled via the 'set led' command
+// can be either the onboard LED on the ESP8266 (LED_BULLETIN) or external via an external pull-up LED
+// (e.g. D1/5 on a bbqkees' board
+// can be enabled and disabled via the 'set led'
 // pin can be set by 'set led_gpio'
 #define EMSESP_LED_GPIO LED_BUILTIN